correct_bugs
correct bugs in parallel tracking processing

+++ tracking_worms/getSegWorm.py
@@ -116,12 +116,23 @@ status_queue = '', base_name = '', min_displacement = 20, thresh_smooth_window =
 def getSegWorm_noMATLABeng(masked_image_file, trajectories_file, segworm_file, \
 status_queue = '', base_name = '', min_displacement = 20, 
 thresh_smooth_window = 1501, csv_tmp_dir = ''):
+
     sendQueueOrPrint(status_queue, 'Obtaining valid trajectories...', base_name);
-    csv_tmp = csv_tmp_dir + 'tmp_' + base_name + '.csv'
-    getValidTrajectories(trajectories_file, min_displacement, thresh_smooth_window, save_csv_name = csv_tmp);
+    csv_tmp = os.path.abspath(csv_tmp_dir + 'tmp_' + base_name + '.csv')
+    
+    df = getValidTrajectories(trajectories_file, min_displacement, thresh_smooth_window, save_csv_name = csv_tmp);
+    if len(df)==0:
+        print "Segworm: no valid trajectories, nothing to do here"
+        return
+
     #calculate segworm data using the MATLAB engine
 #%%
-    cmd = """matlab -nojvm -nodisplay -nosplash -r "addpath(genpath('/Users/ajaver/GitHub_repositories/Multiworm_Tracking/OnlySegWorm/')); movie2segworm_csv('%s', '%s', '%s'); exit;" """  % (csv_tmp, masked_image_file, segworm_file);
+    matlab_path = (os.sep).join([os.path.dirname(os.path.abspath(__file__)), '..', 'OnlySegWorm']) + os.sep;
+    matlab_path = '/Users/ajaver/Documents/GitHub/Multiworm_Tracking/OnlySegWorm/'
+    cmd = """matlab -nojvm -nodisplay -nosplash -r "addpath(genpath('%s')); movie2segworm_csv('%s', '%s', '%s'); exit();" </dev/null """  \
+    % (matlab_path, csv_tmp, masked_image_file, segworm_file);
+
+    print(cmd)
 #%%    
     os.system(cmd)
     os.remove(csv_tmp)
@@ -186,6 +197,6 @@ if __name__ == '__main__':
         except:
             print('%s: Segworm failed' % base_name)
             n_trials +=1;
-            time.sleep(30)
+            time.sleep(1)
             if n_trials == 5:
                 raise 'Segworm failed'
